feat: :sparkles: agregado busqueda y paginacion de modelo

+++ app/Http/Controllers/ModeloController.php
@@ -7,14 +7,34 @@ use Illuminate\Http\Request;
 
 class ModeloController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        $modelos = Modelo::orderBy('id', 'desc')->get();
-        return inertia('Modelos/Index', compact('modelos'));
+        if (!request()->user()->tienePermiso('modelo.listar')) {
+            return redirect()->route('dashboard')->with('error', 'No tenés permiso para listar modelos.');
+        }
+
+        return inertia('Modelos/Index', [
+            'modelos' => Modelo::query()
+            ->when(
+                $request->busqueda,
+                function ($query) use ($request) {
+                    $query->whereRaw('LOWER(nombre) like LOWER(?)', ['%' . $request->busqueda . '%']);
+                }
+            )
+            ->orderBy('id', 'desc')
+            ->paginate(5)
+            ->withQueryString(),
+
+            'terminosBusqueda' => $request->busqueda
+        ]);
     }
 
     public function create()
     {
+        if (!request()->user()->tienePermiso('modelo.crear')) {
+            return redirect()->route('dashboard')->with('error', 'No tenés permiso para crear modelos.');
+        }
+
         return inertia('Modelos/Create');
     }
 
@@ -31,6 +51,10 @@ class ModeloController extends Controller
 
     public function edit(Modelo $modelo)
     {
+        if (!request()->user()->tienePermiso('modelo.editar')) {
+            return redirect()->route('dashboard')->with('error', 'No tenés permiso para editar modelos.');
+        }
+
         return inertia('Modelos/Edit', compact('modelo'));
     }
 
@@ -47,6 +71,10 @@ class ModeloController extends Controller
 
     public function destroy(Modelo $modelo)
     {
+        if ($modelo->motores()->count() > 0) {
+            return redirect()->route('modelos.index')->with('error', 'No se puede eliminar el modelo porque tiene motores asociados.');
+        }
+
         $modelo->delete();
 
         return redirect()->route('modelos.index')->with('success', 'Modelo eliminado correctamente.');
